code/requirements.txt
scikit-learn
pyRserve
numpy
python-csv
pandas
liac-arff